+++ pkg/cloudprovider/providers/anchnet/anchnet_instances.go
@@ -20,13 +20,11 @@ import (
 	"errors"
 	"fmt"
 	"net"
-	"strings"
+	"os"
 	"time"
-	"unicode"
 
 	"github.com/golang/glog"
 	"k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/api/resource"
 	"k8s.io/kubernetes/pkg/cloudprovider"
 
 	anchnet_client "github.com/caicloud/anchnet-go"
@@ -54,42 +52,59 @@ var _ cloudprovider.Instances = (*Anchnet)(nil)
 func (an *Anchnet) NodeAddresses(name string) ([]api.NodeAddress, error) {
 	name = convertToInstanceID(name)
 
+	// Return directly if we find node address in cache.
 	data, exists, err := an.addressCache.GetByKey(name)
 	if exists && err == nil {
 		return data.(AddressCacheEntry).addresses, nil
 	}
 
+	// Can't find address in cache, build it.
+	var addresses []api.NodeAddress
+
+	// Find public IP address.
 	response, err := an.describeInstance(name)
 	if err != nil {
 		return nil, err
 	}
+	addresses = append(addresses, api.NodeAddress{Type: api.NodeExternalIP, Address: response.ItemSet[0].EIP.EipAddr})
 
-	// Here we have some assumptions:
-	// 1. Private SDN is on eth1, and can alwasys be found.
-	// 2. The host calling NodeAddresses matches 'name', i.e. the function is
-	//  called from kubelet, not other controllers. This is true for now, but
-	//  we may need to fix it later, we can use:
-	//  ssh ubuntu@external_ip ifconfig | grep -A 1 'eth1' | tail -1 | cut -d ':' -f 2 | cut -d ' ' -f 1
-	var ip net.IP
-	ifaces, _ := net.Interfaces()
-	for _, iface := range ifaces {
-		if iface.Name == "eth1" {
-			addrs, _ := iface.Addrs()
-			for _, addr := range addrs {
-				switch v := addr.(type) {
-				case *net.IPNet:
-					if v.IP.To4() != nil {
-						ip = v.IP
-						break
+	// Find private IP address (private SDN). Note it's not possbile to find private IP
+	// address for instance other than self, i.e. to find private IP, the function must
+	// be called from kubelet running on instance 'name'. An alternative is to ssh into
+	// the machine and run `ifconfig`, but this requires setting up ssh.
+	// $ ssh ubuntu@external_ip ifconfig | grep -A 1 'eth1' | tail -1 | cut -d ':' -f 2 | cut -d ' ' -f 1
+	searchPrivateIP := true
+	hostname, err := os.Hostname()
+	if err != nil {
+		glog.Warningf("unable to get hostname %v", err)
+		searchPrivateIP = false
+	}
+	instanceID := convertToInstanceID(hostname)
+	if instanceID != name {
+		glog.Warningf("unable to find node addresses for instance other than self instance")
+		searchPrivateIP = false
+	}
+
+	// Here we assume private SDN is on eth1, and can alwasys be found.
+	// TODO: Need a reliable way to get interface name.
+	if searchPrivateIP {
+		var ip net.IP
+		ifaces, _ := net.Interfaces()
+		for _, iface := range ifaces {
+			if iface.Name == "eth1" {
+				addrs, _ := iface.Addrs()
+				for _, addr := range addrs {
+					switch v := addr.(type) {
+					case *net.IPNet:
+						if v.IP.To4() != nil {
+							ip = v.IP
+							break
+						}
 					}
 				}
 			}
 		}
-	}
-
-	addresses := []api.NodeAddress{
-		{Type: api.NodeInternalIP, Address: ip.String()},
-		{Type: api.NodeExternalIP, Address: response.ItemSet[0].EIP.EipAddr},
+		addresses = append(addresses, api.NodeAddress{Type: api.NodeInternalIP, Address: ip.String()})
 	}
 
 	an.addressCache.Add(AddressCacheEntry{
@@ -195,23 +210,4 @@ func (an *Anchnet) searchInstances(search_word string) (*anchnet_client.Describe
 		time.Sleep(RetryIntervalOnError)
 	}
 	return nil, fmt.Errorf("Unable to find instance with name %v\n", search_word)
-}
-
-// makeResources converts bare resources to api spec'd resource, cpu is in cores, memory is in GiB.
-func makeResources(cpu, memory int) *api.NodeResources {
-	return &api.NodeResources{
-		Capacity: api.ResourceList{
-			api.ResourceCPU:    *resource.NewMilliQuantity(int64(cpu*1000), resource.DecimalSI),
-			api.ResourceMemory: *resource.NewQuantity(int64(memory*1024*1024*1024), resource.BinarySI),
-		},
-	}
-}
-
-// convertToInstanceID converts name to anchnet instance ID, e.g.
-//   i-ff830wku->i-FF830WKU, i-FF830WKU->i-FF830WKU.
-func convertToInstanceID(name string) string {
-	s := strings.ToUpper(name)
-	a := []rune(s)
-	a[0] = unicode.ToLower(a[0])
-	return string(a)
 }